meta: Move Base64 decoder into utils.
